Adjusted addresses used in PRSwitchGetStates to reflect FPGA change.  Use the combined version/revision to choose between the old address and the new one so that old FPGAs will still work fine.

+++ src/PRDevice.h
@@ -140,7 +140,7 @@ protected:
 
     uint16_t version;
     uint16_t revision;
-    uint16_t combinedVersionRevision;
+    uint32_t combinedVersionRevision;
     /**
      * Calculated combined Version/Revision number.
      */
